feat!: adopt brand-based NonEmptyArray; bump to ts-type-forge v7 / ts-data-forge v11
Adopt the brand-based `NonEmptyArray` from ts-type-forge / ts-data-forge and
tidy up the length-constrained array validators.

- Bump `ts-type-forge` to ^7.0.0 and `ts-data-forge` to ^11.0.0.
- Schema composers (`union`, `intersection`, `mergeRecords`, `partial`,
  `required`, `enumType`) take structural `NonEmptyTuple` inputs, so a plain
  non-empty tuple of schemas / keys is still accepted.
- `nonEmptyArray` now produces the branded `NonEmptyArray` output, built on the
  shared array-`Type` builder.
- Split `length-constrained-array.mts` into one file per validator
  (`fixed`/`min`/`max`/`bounded-length-array.mts`), mirroring the tuple family;
  tests split to match.
- Move the shared `buildType` / `validateElements` helpers into an internal
  `build-type.mts` that is excluded from index generation, so they are no
  longer part of the public package API. `buildType` is generic over the
  branded output type, so each validator returns its branded array type without
  a call-site assertion.
- `combination.test`: build `paletteBase` with `fixedLengthArray` so the
  intersection default is assignable directly, removing an `as unknown as` cast.
- Make TypeDoc source links deterministic via `sourceLinkTemplate`, and trim the
  verbose v10.0.0 CHANGELOG entry.

BREAKING CHANGE: `nonEmptyArray(elementType)` now yields the brand-based
`NonEmptyArray` value type (like the other branded array validators); a branded
non-empty array is obtained through validation / cast rather than a plain array
literal. Requires `ts-type-forge@^7.0.0` and `ts-data-forge@^11.0.0`.
