+++ src/array/index.mts
@@ -1,8 +1,11 @@
 export * from './array.mjs';
+export * from './bounded-length-array.mjs';
 export * from './bounded-length-tuple.mjs';
+export * from './fixed-length-array.mjs';
 export * from './fixed-length-tuple.mjs';
-export * from './length-constrained-array.mjs';
+export * from './max-length-array.mjs';
 export * from './max-length-tuple.mjs';
+export * from './min-length-array.mjs';
 export * from './min-length-tuple.mjs';
 export * from './non-empty-array.mjs';
 export * from './tuple.mjs';
