+++ test/combination.test.mts
@@ -4,7 +4,7 @@ import {
   array,
   brandedString,
   enumType,
-  fixedLengthTuple,
+  fixedLengthArray,
   int,
   intersection,
   intRange,
@@ -280,7 +280,11 @@ describe('advanced type', () => {
     intRange({ start: -128, end: 128, defaultValue: 0 }),
   ]);
 
-  const paletteBase = fixedLengthTuple(2, EvenRange);
+  // A fixed-length-2 branded array is already non-empty, so
+  // `NonEmptyArray<EvenRange> & FixedLengthArray<2, EvenRange>` collapses to
+  // `FixedLengthArray<2, EvenRange>`; `paletteBase` therefore satisfies the
+  // intersection default directly, with no unsafe cast.
+  const paletteBase = fixedLengthArray(2, EvenRange);
 
   const Palette = intersection(
     [nonEmptyArray(EvenRange), paletteBase],
@@ -415,11 +419,11 @@ describe('advanced type', () => {
     assert.deepStrictEqual(messages, [
       'Error at status: expected one of <enum>, <null>, <undefined> but <string> type value "unknown" was passed.',
       'Error at coordinates.1: expected an integer between -128 and 127 but `190` was passed.',
-      'Error at palette: expected value to match all types of <NonEmptyArray<EvenRange>>, <FixedLengthTuple<2, EvenRange>> but <object> type value `[1,3,5]` was passed.',
+      'Error at palette: expected value to match all types of <NonEmptyArray<EvenRange>>, <FixedLengthArray<2, EvenRange>> but <object> type value `[1,3,5]` was passed.',
       'Error at palette.0: expected <EvenRange> type but <number> type value `1` was passed.',
       'Error at palette.1: expected <EvenRange> type but <number> type value `3` was passed.',
       'Error at palette.2: expected <EvenRange> type but <number> type value `5` was passed.',
-      'Error at palette: expected value to match all types of <NonEmptyArray<EvenRange>>, <FixedLengthTuple<2, EvenRange>> but <object> type value `[1,3,5]` was passed.',
+      'Error at palette: expected value to match all types of <NonEmptyArray<EvenRange>>, <FixedLengthArray<2, EvenRange>> but <object> type value `[1,3,5]` was passed.',
       'Error at palette: expected array of length 2 but length 3 was passed.',
       'Error at metrics: expected one of <key-value-record>, <undefined> but <object> type value `{"alpha":1}` was passed.',
       'Error at tags: expected one of <NonEmptyArray<Tag>>, <undefined> but <object> type value `[]` was passed.',
@@ -434,7 +438,7 @@ describe('advanced type', () => {
       status: undefined,
     } as const;
 
-    assert.deepStrictEqual(AdvancedNodeType.fill(partialNode), {
+    assert.deepStrictEqual<UnknownRecord>(AdvancedNodeType.fill(partialNode), {
       id: Identifier.cast('id:partial'),
       status: undefined,
       coordinates: [0, 0],
